Explicitly call name() on the Enum type instead of the implementation. (#98)
This saves one method reference per enum type used in an parcelable auto value type. The reason for this is that casts to supertypes are free (that is, they don't actually emit a cast). Since name() is a method on Enum and final, we can safely reference it directly using the superclass cast.